requirements.txt
Django>=2.2.9
pytz==2019.3
sqlparse==0.3.0
djangorestframework
django-rest-auth
django-allauth
asgiref==3.2.3
django-cors-headers
postgres
psycopg2-binary
drf-yasg
swagger_spec_validator
